Addressing code review feedback
